chore: remove bottom sheet from invalid mdl
